MODIFY : models.py 수정요청사항반영

+++ products/models.py
@@ -5,6 +5,7 @@ class Category(models.Model):
 
     class Meta:
         db_table = 'categories'
+        
     def __str__(self):
         return self.name
 
@@ -14,6 +15,7 @@ class SubCategory(models.Model):
 
     class Meta:
         db_table = 'sub_categories'
+
     def __str__(self):
         return self.name
 
@@ -22,6 +24,7 @@ class Brand(models.Model):
 
     class Meta:
         db_table = 'brands'
+
     def __str__(self):
         return self.name
 
@@ -30,6 +33,7 @@ class Color(models.Model):
 
     class Meta:
         db_table = 'colors'
+
     def __str__(self):
         return self.name
 
@@ -41,13 +45,14 @@ class Furniture(models.Model):
 
     class Meta:
         db_table = 'furnitures'
+
     def __str__(self):
         return self.name
 
 class Product(models.Model):
     thumbnail_image_url = models.CharField(max_length=200)
     main_image_url      = models.CharField(max_length=200)
-    price               = models.IntegerField()
+    price               = models.DecimalField(max_digits=11, decimal_places=2)
     is_new              = models.BooleanField()
     sub_category        = models.ForeignKey(SubCategory, on_delete=models.CASCADE)
     color               = models.ForeignKey(Color, on_delete=models.CASCADE)
